API Reference

GadgetUnits.GadgetPhysical

GadgetUnits.GadgetPhysical

GadgetUnits.GadgetPhysicalUnits

GadgetUnits.GadgetPhysicalUnits

Cosmology.age

Cosmology.cosmology

GadgetUnits.arcmin_to_kpc

GadgetUnits.arcmin_to_kpc

GadgetUnits.arcmin_to_kpc

GadgetUnits.arcmin_to_kpc

GadgetUnits.mJy_to_W

GadgetUnits.mJy_to_W

GadgetUnits.mJy_to_W

Exported Functions

Cosmology.age — Function

age(h::SnapshotHeader)

Computes the age of the universe given the properties from SnapshotHeader.

Cosmology.cosmology — Method

cosmology(h::SnapshotHeader)

Defines a Cosmology.AbstractCosmology from the properties of h.

GadgetUnits.arcmin_to_kpc — Method

arcmin_to_kpc(c::Cosmology.AbstractCosmology, θ::Real, z::Real)

Convert arcmin to kpc for a given redshift z and cosmology c. Deletes unit information.

GadgetUnits.arcmin_to_kpc — Method

arcmin_to_kpc(c::Cosmology.AbstractCosmology, θ::Unitful.AbstractQuantity, z::Real)

Convert arcmin to kpc for a given redshift z and cosmology c. Conserves unit information.

GadgetUnits.arcmin_to_kpc — Method

arcmin_to_kpc(θ::Real, h::SnapshotHeader)

Convert arcmin to kpc for a given redshift and cosmology taken from SnapshotHeader.

GadgetUnits.arcmin_to_kpc — Method

arcmin_to_kpc(θ::Unitful.AbstractQuantity, h::SnapshotHeader)

Convert arcmin to kpc for a given redshift and cosmology taken from SnapshotHeader.

GadgetUnits.mJy_to_W — Method

mJy_to_W( c::Cosmology.AbstractCosmology, S::Real, z::Real )

Converts synchrotron emission S in [mJy] to [W/Hz] for a given redshift z and cosmology c. Deletes unit information.

GadgetUnits.mJy_to_W — Method

mJy_to_W( c::Cosmology.AbstractCosmology, S::Unitful.AbstractQuantity, z::Real )

Converts synchrotron emission S in [mJy] to [W/Hz] for a given redshift z and cosmology c. Conserves unit information.

GadgetUnits.mJy_to_W — Method

mJy_to_W(S::Union{Real, Unitful.AbstractQuantity}, h::SnapshotHeader)

Converts synchrotron emission S in [mJy] to [W/Hz] for a given redshift and cosmology taken from SnapshotHeader.

Exported Types

GadgetUnits.GadgetPhysical — Type

GadgetPhysical( DT::DataType, 
                l_unit::Real=3.085678e21, m_unit::Real=1.989e43, v_unit::Real=1.e5;
                a_scale::Real=1.0, hpar::Real=1.0,
                γ_th::Real=5.0/3.0, γ_CR::Real=4.0/3.0, xH::Real=0.76)

Set up a unit struct with a given DataType.

GadgetUnits.GadgetPhysical — Method

GadgetPhysical(l_unit::T=3.085678e21, m_unit::T=1.989e43, v_unit::T=1.e5;
            a_scale::T=1.0, hpar::T=1.0,
            γ_th::T=5.0/3.0, γ_CR::T=4.0/3.0, xH::T=0.76) where T

Creates a datatype GadgetPhysical which holds the conversion factors between comoving code units and physical units, without unit information.

Examples

julia> GU = GadgetPhysical()
GadgetPhysicalUnits(3.085678e21, 100000.0, 1.989e43, 3.085678e16, 977.7923542981722, 1.989e53, 1.2414361549102458e65, 1.0, 6.769911178294544e-22, 743179.9340255889, 47.50882854026919, 6.769911178294544e-12, 6.769911178294544e-12)
julia> GU.x_cgs
3.085678e21

Keyword arugments specify:

Arguments

a_scale::T = 1.0: Cosmological scale factor of the simulation. Can be passed with the header h as h.time.

hpar::T = 1.0: Hubble constant as 'little h'. Can be passed with header h as h.h0.

γ_th::T = 5.0/3.0: Adiabatic index of gas.

γ_CR::T = 4.0/3.0: Adiabatic index of cosmic ray component.

xH::T = 0.76: Hydrogen fraction of the simulation, if run without chemical model.

Fields

| Name | Meaning | |: –––––––––––- |:––––––––––––––- | | x_cgs::T | position in cm | | x_kpc::T | position in kpc | | v_cgs::T | velocity in cm/s | | v_kms::T | velocity in km/s | | m_cgs::T | mass in g | | m_msun::T | mass in Msun | | t_s::T | time in sec | | t_Myr::T | time in Myr | | E_cgs::T | energy in erg | | E_eV::T | energy in eV | | B_cgs::T | magnetic field in Gauss | | rho_cgs::T | density in $g/cm^3$ | | rho_ncm3::T | density in $n_p/cm^3$ | | T_K::T | temperature in K | | T_eV::T | temperature in eV | | P_th_cgs::T | thermal pressure in Ba | | P_CR_cgs::T | cosmic ray pressure in Ba |

GadgetUnits.GadgetPhysicalUnits — Type

GadgetPhysicalUnits( DT::DataType, 
                        l_unit::Real=3.085678e21, m_unit::Real=1.989e43, v_unit::Real=1.e5;
                        a_scale::Real=1.0, hpar::Real=1.0,
                        γ_th::Real=5.0/3.0, γ_CR::Real=4.0/3.0, xH::Real=0.76)

Set up a unit struct with a given DataType.

GadgetUnits.GadgetPhysicalUnits — Method

GadgetPhysicalUnits(l_unit::T=3.085678e21, m_unit::T=1.989e43, v_unit::T=1.e5;
                    a_scale::T=1.0, hpar::T=1.0,
                    γ_th::T=5.0/3.0, γ_CR::T=4.0/3.0, xH::T=0.76) where T

Creates a datatype GadgetPhysicalUnits which holds the conversion factors between comoving code units and physical units. Stores the unit information which can be converted with Unitful or UnitfulAstro.

Examples

julia> GU = GadgetPhysicalUnits()
GadgetPhysicalUnits(3.085678e21 cm, 100000.0 cm s^-1, 1.989e43 g, 3.085678e16 s, 977.7923542981722 Myr, 1.989e53 erg, 1.2414361549102458e65 eV, 1.0 Gs, 6.769911178294544e-22 g cm^-3, 743179.9340255889 N_e/cm^3, 47.50882854026919 K, 6.769911178294544e-12 Ba, 6.769911178294544e-12 Ba)
julia> GU.x_cgs
3.085678e21 cm

Keyword Arguments

a_scale::T = 1.0: Cosmological scale factor of the simulation. Can be passed with the header h as h.time.

hpar::T = 1.0: Hubble constant as 'little h'. Can be passed with header h as h.h0.

γ_th::T = 5.0/3.0: Adiabatic index of gas.

γ_CR::T = 4.0/3.0: Adiabatic index of cosmic ray component.

xH::T = 0.76: Hydrogen fraction of the simulation, if run without chemical model.

Fields

| Name | Meaning | |: –––––––––––––- |:––––––––––––––- | | x_cgs::Quantity{T} | position in cm | | x_kpc::Quantity{T} | position in kpc | | v_cgs::Quantity{T} | velocity in cm/s | | v_kms::Quantity{T} | velocity in km/s | | m_cgs::Quantity{T} | mass in g | | m_msun::Quantity{T} | mass in Msun | | t_s::Quantity{T} | time in sec | | t_Myr::Quantity{T} | time in Myr | | E_cgs::Quantity{T} | energy in erg | | E_eV::Quantity{T} | energy in eV | | B_cgs::Quantity{T} | magnetic field in Gauss | | rho_cgs::Quantity{T} | density in $g/cm^3$ | | rho_ncm3::Quantity{T} | density in $n_p/cm^3$ | | T_K::Quantity{T} | temperature in K | | T_eV::Quantity{T} | temperature in eV | | P_th_cgs::Quantity{T} | thermal pressure in Ba | | P_CR_cgs::Quantity{T} | cosmic ray pressure in Ba |

Private Functions

Private Types