Conversion Structs

GadgetUnits.jl uses Unitful.jl and UnitfulAstro.jl to store the unit conversion factors with actual units in place. You can convert the internal units of Gadget into cgs units by defining the struct

GadgetUnits.GadgetPhysicalUnits — Type

GadgetPhysicalUnits(l_unit::T=3.085678e21, m_unit::T=1.989e43, v_unit::T=1.e5;
                    a_scale::T=1.0, hpar::T=1.0,
                    γ_th::T=5.0/3.0, xH::T=0.752) where T

Creates a struct GadgetPhysicalUnits which holds the conversion factors between comoving code units and physical units. Stores the unit information which can be converted with Unitful or UnitfulAstro.

Keyword Arguments

a_scale::T = 1.0: Cosmological scale factor of the simulation. Can be passed with the header h as h.time.

hpar::T = 1.0: Hubble constant as 'little h'. Can be passed with header h as h.h0.

γ_th::T = 5.0/3.0: Adiabatic index of gas.

xH::T = 0.752: Hydrogen fraction of the simulation, if run without chemical model.

Fields

Name	Meaning

x_cgs::Quantity{T}	position in cm

x_physical::Quantity{T}	position in kpc

v_cgs::Quantity{T}	velocity in cm/s

v_physical::Quantity{T}	velocity in km/s

m_cgs::Quantity{T}	mass in g

m_msun::Quantity{T}	mass in solar masses

t_s::Quantity{T}	time in sec

t_Myr::Quantity{T}	time in Myr

E_cgs::Quantity{T}	energy in erg

E_eV::Quantity{T}	energy in eV

B_cgs::Quantity{T}	magnetic field in Gauss

rho_cgs::Quantity{T}	density in $g/cm^3$

rho_ncm3::Quantity{T}	density in $n_p/cm^3$

T_K::Quantity{T}	temperature in K

T_eV::Quantity{T}	temperature in eV

P_th_cgs::Quantity{T}	thermal pressure in Ba

P_CR_cgs::Quantity{T}	cosmic ray pressure in Ba

CR_norm::Quantity{T}	LMBSPECTRALCRs Norm in cgs units

GadgetPhysicalUnits( DT::DataType, 
                        l_unit::Real=3.085678e21, m_unit::Real=1.989e43, v_unit::Real=1.e5;
                        a_scale::Real=1.0, hpar::Real=1.0,
                        γ_th::Real=5.0/3.0, xH::Real=0.752)

Set up a unit struct with a given DataType.

GadgetPhysical( h::SnapshotHeader, 
                l_unit::Real=3.085678e21, m_unit::Real=1.989e43, v_unit::Real=1.e5;
                γ_th::Real=5.0/3.0, xH::Real=0.752)

Set up a unit struct with a given SnapshotHeader. Only works in 64-bits.

If you want to have the same functionality, but without using Unitful.jl you can construct a similar struct:

GadgetUnits.GadgetPhysical — Type

GadgetPhysical(l_unit::T=3.085678e21, m_unit::T=1.989e43, v_unit::T=1.e5;
               a_scale::T=1.0, hpar::T=1.0,
               γ_th::T=5.0/3.0, xH::T=0.752) where T

Creates a struct GadgetPhysical which holds the conversion factors between comoving code units and physical units, without unit information.

Keyword arugments specify:

Arguments

a_scale::T = 1.0: Cosmological scale factor of the simulation. Can be passed with the header h as h.time.

hpar::T = 1.0: Hubble constant as 'little h'. Can be passed with header h as h.h0.

γ_th::T = 5.0/3.0: Adiabatic index of gas.

xH::T = 0.752: Hydrogen fraction of the simulation, if run without chemical model.

Fields

Name	Meaning

x_cgs::T	position in cm

x_physical::T	position in kpc

v_cgs::T	velocity in cm/s

v_physical::T	velocity in km/s

m_cgs::T	mass in g

m_msun::T	mass in Msun

m_physical::T	mass in 10^10 Msun

t_s::T	time in sec

t_Myr::T	time in Myr

E_cgs::T	energy in erg

E_eV::T	energy in eV

B_cgs::T	magnetic field in Gauss

rho_physical::T	density in 10^10 Msun/kpc^3

rho_cgs::T	density in $g/cm^3$

rho_ncm3::T	density in $n_p/cm^3$

T_K::T	temperature in K

T_eV::T	temperature in eV

P_th_cgs::T	thermal pressure in Ba

P_CR_cgs::T	cosmic ray pressure in Ba

CR_norm::T	LMBSPECTRALCRs Norm in cgs units

GadgetPhysical( DT::DataType, 
                l_unit::Real=3.085678e21, m_unit::Real=1.989e43, v_unit::Real=1.e5;
                a_scale::Real=1.0, hpar::Real=1.0,
                γ_th::Real=5.0/3.0, xH::Real=0.752)

Set up a unit struct with a given DataType.

GadgetPhysical( h::SnapshotHeader, 
                l_unit::Real=3.085678e21, m_unit::Real=1.989e43, v_unit::Real=1.e5;
                γ_th::Real=5.0/3.0, xH::Real=0.752)

Set up a unit struct with a given SnapshotHeader. Only works in 64-bits.

This uses the same conversions, but leaves out the actual unit properties.

Usage

To convert, say positions of gas particles from a cosmological simulation to physical units you can use:

h     = read_header(filename)
pos   = read_snap(filename, "POS", 0)
GU    = GadgetPhysicalUnits(a_scale=h.time, hpar=h.h0)
pos .*= GU.x_cgs

If you have different units than the standard Gadget ones you can call the cunstructor with different values

GU = GadgetPhysicalUnits(your_l_unit, your_m_unit, your_v_unit; kwargs...)

Converting the units can then be done with Unitful.jl and UnitfulAstro.jl. So if you want to convert the position units from the default cm to Mpc you can do this as:

using Unitful
using UnitfulAstro

pos = read_snap(filename, "POS", 0)
pos = @. pos * GU.x_cgs |> u"Mpc"

If you want to get rid of the units, for example if you need basic datatypes again for a function you can use the funtion ustrip:

pos = ustrip(pos)

Data Types

You can define the datatype of the struct by passing it as an optional first parameter

GadgetUnits.GadgetPhysicalUnits — Method

GadgetPhysicalUnits(l_unit::T=3.085678e21, m_unit::T=1.989e43, v_unit::T=1.e5;
                    a_scale::T=1.0, hpar::T=1.0,
                    γ_th::T=5.0/3.0, xH::T=0.752) where T

Creates a struct GadgetPhysicalUnits which holds the conversion factors between comoving code units and physical units. Stores the unit information which can be converted with Unitful or UnitfulAstro.

Keyword Arguments

a_scale::T = 1.0: Cosmological scale factor of the simulation. Can be passed with the header h as h.time.

hpar::T = 1.0: Hubble constant as 'little h'. Can be passed with header h as h.h0.

γ_th::T = 5.0/3.0: Adiabatic index of gas.

xH::T = 0.752: Hydrogen fraction of the simulation, if run without chemical model.

Fields

Name	Meaning

x_cgs::Quantity{T}	position in cm

x_physical::Quantity{T}	position in kpc

v_cgs::Quantity{T}	velocity in cm/s

v_physical::Quantity{T}	velocity in km/s

m_cgs::Quantity{T}	mass in g

m_msun::Quantity{T}	mass in solar masses

t_s::Quantity{T}	time in sec

t_Myr::Quantity{T}	time in Myr

E_cgs::Quantity{T}	energy in erg

E_eV::Quantity{T}	energy in eV

B_cgs::Quantity{T}	magnetic field in Gauss

rho_cgs::Quantity{T}	density in $g/cm^3$

rho_ncm3::Quantity{T}	density in $n_p/cm^3$

T_K::Quantity{T}	temperature in K

T_eV::Quantity{T}	temperature in eV

P_th_cgs::Quantity{T}	thermal pressure in Ba

P_CR_cgs::Quantity{T}	cosmic ray pressure in Ba

CR_norm::Quantity{T}	LMBSPECTRALCRs Norm in cgs units

GadgetPhysicalUnits( DT::DataType, 
                        l_unit::Real=3.085678e21, m_unit::Real=1.989e43, v_unit::Real=1.e5;
                        a_scale::Real=1.0, hpar::Real=1.0,
                        γ_th::Real=5.0/3.0, xH::Real=0.752)

Set up a unit struct with a given DataType.

DT defaults to Float64.

Please be aware that some unit conversions overflow at Float32!